Major refactor in the whole project: functions have now been split between different files in order to increase readability and to help optimize functions during development.
errorcodes.h : it contains only the structure and macros for ErrorCode type

utility.c/utility.h : it contains auxiliary functions for printing to terminal and for handling backups

memory.c/memory.h : it contains functions that will help us read and write data in a specified area of memory for a given pid

elfscan.c/elfscan.h : it's a mix of functions capable of parsing an elf file. In some cases we use libelf and in others we simply parse the result from objdump commands. These functions will help us get information for a given program such as : PID, elf type, address of start sections (will help evaluate base address if elf type is DYN), address of a function, size of a function, address of a libc function

breakpoint.c : only contains the function that sets a breakpoint at a given address for a given pid -> accordingly to challenge 1

indirectcall.c : only contains the functions that triggers an indirect call with parameters passed as value or reference

interceptor.c : only contains the main and all the function calls that we need in order to intercept a given program

README.md : typo fix